pyproject.toml
[tool.poetry]
name = "imhex-mcp"
version = "1.0.0"
description = "High-performance MCP server for ImHex hex editor"
authors = ["ImHex MCP Team"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
mcp = ">=1.0.0"
pydantic = ">=2.0.0"
zstandard = ">=0.22.0"
pyyaml = ">=6.0.0"
prometheus-client = ">=0.20.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
ruff = "^0.3.0"
isort = "^5.13.0"
mypy = "^1.8.0"
sphinx = "^8.0.0"
sphinx-rtd-theme = "^2.0.0"
sphinxcontrib-mermaid = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["lib", "mcp-server", "tests"]
asyncio_mode = "auto"
